backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pydantic-settings==2.1.0
loguru==0.7.2
httpx==0.27.0
python-multipart==0.0.9
email-validator==2.1.1
pytest==8.1.1
pytest-asyncio==0.23.5
anyio==4.3.0
